Use README.md in the folders (#14)
* Remove rfc numbers

Manually created numbers aren't really helpful. My very first RFC already collided. Things will get merged in a different order. Some PRs will not even make it. So we will end up with lots of holes in the numbering anyway. Let's try this without numbers. It makes creating RFCs even easier and this was the main goal.

* Use README.md in the folders

If we use README.md in the folder then Github will show a preview in the interface. And each rfc directory will look the same (easier to build automation later on).

* remove number from first rfc

* rename first rfc description to README.md